+++ cmd/hidden_lake/adapters/common/recv/main.go
@@ -9,8 +9,7 @@ import (
 	"time"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/adapters/common"
-	"github.com/number571/go-peer/internal/msgconv"
-	"github.com/number571/go-peer/pkg/client/message"
+	net_message "github.com/number571/go-peer/pkg/network/message"
 	"github.com/number571/go-peer/pkg/storage"
 	"github.com/number571/go-peer/pkg/storage/database"
 
@@ -121,7 +120,7 @@ func transferTraffic(db database.IKVDatabase, portService, portHLT int, hasLog b
 	}
 }
 
-func loadMessageFromService(portService int, id uint64) (message.IMessage, error) {
+func loadMessageFromService(portService int, id uint64) (net_message.IMessage, error) {
 	// build request to service
 	req, err := http.NewRequest(
 		http.MethodGet,
@@ -150,8 +149,10 @@ func loadMessageFromService(portService int, id uint64) (message.IMessage, error
 		return nil, fmt.Errorf("failed: incorrect response from service")
 	}
 
-	msgString := string(msgStringAsBytes[1:])
-	msg := message.LoadMessage(getMessageSettings(), msgconv.FromStringToBytes(msgString))
+	msg := net_message.LoadMessage(
+		getMessageSettings(),
+		string(msgStringAsBytes[1:]),
+	)
 	if msg == nil {
 		return nil, fmt.Errorf("message is nil")
 	}
@@ -224,9 +225,8 @@ func incrementCountInDB(db database.IKVDatabase) error {
 	return nil
 }
 
-func getMessageSettings() message.ISettings {
-	return message.NewSettings(&message.SSettings{
-		FWorkSizeBits:     workSize,
-		FMessageSizeBytes: messageSize,
+func getMessageSettings() net_message.ISettings {
+	return net_message.NewSettings(&net_message.SSettings{
+		FWorkSizeBits: workSize,
 	})
 }