+++ cmd/hidden_lake/traffic/internal/database/types.go
@@ -1,30 +1,30 @@
 package database
 
 import (
-	"github.com/number571/go-peer/pkg/client/message"
+	net_message "github.com/number571/go-peer/pkg/network/message"
 	"github.com/number571/go-peer/pkg/types"
 )
 
-type IKVDatabase interface {
+type IDatabase interface {
 	types.ICloser
 
-	Push(message.IMessage) error
-	Load([]byte) (message.IMessage, error)
+	Push(net_message.IMessage) error
+	Load([]byte) (net_message.IMessage, error)
 	Hashes() ([][]byte, error)
 
 	Settings() ISettings
 }
 
 type ISettings interface {
+	net_message.ISettings
+
 	GetPath() string
 	GetMessagesCapacity() uint64
-	GetMessageSizeBytes() uint64
-	GetWorkSizeBits() uint64
 }
 
 type IWrapperDB interface {
 	types.ICloser
 
-	Get() IKVDatabase
-	Set(IKVDatabase) IWrapperDB
+	Get() IDatabase
+	Set(IDatabase) IWrapperDB
 }