+++ cmd/hidden_lake/traffic/pkg/app/init_database.go
@@ -12,17 +12,17 @@ import (
 func (p *sApp) initDatabase() error {
 	sett := database.NewSettings(&database.SSettings{
 		FPath:             fmt.Sprintf("%s/%s", p.fPathTo, hlt_settings.CPathDB),
+		FNetworkKey:       p.fConfig.GetNetworkKey(),
 		FMessagesCapacity: p.fConfig.GetSettings().GetMessagesCapacity(),
-		FMessageSizeBytes: p.fConfig.GetSettings().GetMessageSizeBytes(),
 		FWorkSizeBits:     p.fConfig.GetSettings().GetWorkSizeBits(),
 	})
 
 	if !p.fConfig.GetIsStorage() {
-		p.fWrapperDB.Set(database.NewVoidKeyValueDB(sett))
+		p.fWrapperDB.Set(database.NewVoidDatabase(sett))
 		return nil
 	}
 
-	db, err := database.NewKeyValueDB(sett)
+	db, err := database.NewDatabase(sett)
 	if err != nil {
 		return errors.WrapError(err, "init database")
 	}