+++ pkg/queue_set/types.go
@@ -2,6 +2,7 @@ package queue_set
 
 type IQueueSet interface {
 	GetSettings() ISettings
+	GetQueueKeys() [][]byte
 
 	IQueuePusher
 	IQueueLoader